fix(backend-metal): BatchNorm train ループのカウンタを ulong 化し wraparound を防ぐ
`batch_norm_train_f32` のチャネル走査ループ（`ch += grid_size`）・
要素走査ループ（`i += BATCH_NORM_SIMD_WIDTH`。パス 1〜3 の 3 箇所）は
`uint`（32-bit）カウンタのまま加算していたため、起動前検証が受理する
`m=u32::MAX` 近傍の形状で加算が折り返し、ループが終了しない
（GPU ハング）不具合があった。カウンタを `ulong` へ変更し、境界比較
（`ch < c`・`i < m`）も `ulong` 側で行うことで折り返しを構造的に排除
する（codex-review P1・Cursor Bugbot Medium 指摘）。

あわせて文字列証跡テスト（`batch_norm_source_evidence.rs`）を新カーネル
本文へ追従させ、`ulong` カウンタの回帰を検出する新規テストを追加した。

その他 codex-review P2 指摘のうち対応可能なものを修正した:
- `batch_norm_train_extreme_values_stay_finite`（Metal 実機限定
  `#[ignore]` テスト）の入力値を `2e20` から `1e18` 規模へ調整した。
  内部 soft-f64（`f64` 相当）計算自体は `2e20` 級でも有限のまま
  完結するが、最終的に `var` を `f32` へ narrow する時点で
  `f32::MAX` を超え正当に `+inf` になりうるため、本テストの意図
  （soft-f64 機構が NaN／中間 overflow を起こさないことの確認）に
  沿わない値だった。
- 空軸早期 return テストで `mean`／`var` に空スライスを渡していたため
  `validate_batch_norm_launch` が `n==0` の早期 return より先に長さ
  `c` 不一致で `InvalidBatchNormShape` を返していた（Cursor Bugbot
  Low 指摘と同一箇所）。CPU 側の同型テストに合わせ長さ `c` のダミー
  値を渡すよう修正した。
- `docs/batch-norm-ops-design.md` §9.6 が
  `crates/facade/tests/batch_norm_backend_parity.rs` へ Metal 版
  infer／backward／rank4 テストを本 issue で追加したと誤記していた
  （実際に存在するのは兄弟 issue #1735 由来の CUDA 版のみ）ため訂正し、
  §9.7 へ対象外として引き継いだ。

なお、`batch_norm_model.rs` のホストモデル縮約順序（索引順逐次加算）が
カーネルの 32 レーン butterfly 縮約と異なり、極端に桁の離れた入力で
両者が乖離しうる点（codex-review P2）は、ホストモデル側で butterfly
縮約を厳密に再現する大掛かりな書き換えを要するためスコープ外とし、
PR 本文の対象外節へ記録した。

Refs #1736

+++ crates/backend-metal/src/shaders/batch_norm.metal
@@ -897,10 +897,10 @@ kernel void batch_norm_train_f32(
 
     ulong m_f64 = (((ulong)m_f64_hi) << 32) | (ulong)m_f64_lo;
 
-    for (uint ch = tg_id; ch < c; ch += grid_size) {
+    for (ulong ch = (ulong)tg_id; ch < (ulong)c; ch += (ulong)grid_size) {
         // パス 1: 平均（soft-f64 総和。冒頭コメント「総和の順序」参照）。
         ulong lane_sum = 0ul; // +0.0（f64）。
-        for (uint i = lane; i < m; i += BATCH_NORM_SIMD_WIDTH) {
+        for (ulong i = (ulong)lane; i < (ulong)m; i += (ulong)BATCH_NORM_SIMD_WIDTH) {
             ulong xv = bn_f64_widen(as_type<uint>(x[BN_IDX(i)]));
             lane_sum = bn_f64_add(lane_sum, xv);
         }
@@ -914,7 +914,7 @@ kernel void batch_norm_train_f32(
 
         // パス 2: 分散（二パス。`(x-mean)^2` を soft-f64 で蓄積する）。
         ulong lane_sq = 0ul;
-        for (uint i = lane; i < m; i += BATCH_NORM_SIMD_WIDTH) {
+        for (ulong i = (ulong)lane; i < (ulong)m; i += (ulong)BATCH_NORM_SIMD_WIDTH) {
             ulong xv = bn_f64_widen(as_type<uint>(x[BN_IDX(i)]));
             ulong dev = bn_f64_sub(xv, mean);
             ulong devsq = bn_f64_mul(dev, dev);
@@ -944,7 +944,7 @@ kernel void batch_norm_train_f32(
         float bv = (has_bias != 0) ? b[ch] : 0.0f;
         ulong wv64 = bn_f64_widen(as_type<uint>(wv));
         ulong bv64 = bn_f64_widen(as_type<uint>(bv));
-        for (uint i = lane; i < m; i += BATCH_NORM_SIMD_WIDTH) {
+        for (ulong i = (ulong)lane; i < (ulong)m; i += (ulong)BATCH_NORM_SIMD_WIDTH) {
             ulong idx = BN_IDX(i);
             ulong xv = bn_f64_widen(as_type<uint>(x[idx]));
             ulong dev = bn_f64_sub(xv, mean);
